attrib: FindCollectionWithDefault's key was truncated -- the same bug, again (G5)
I went at the Attrib gap G5 stub expecting a subsystem and found a signature bug
instead, of exactly the kind attrib_findcollection.h already documents fixing
for the sibling function two declarations above.

FindCollectionWithDefault was declared `(int liKey)`, with a note claiming the
callee "reads only the low word 0x85B5C4F4". Its only call site says otherwise --
Attrib::Gen::surfacelist::ChangeWithDefault @0x8227EFD8-EC:

    lis  r11, -0x7A4B ; ori r3, r11, 0xC4F4   -> r3  = 0x85B5C4F4
    lis  r11, 0x42C2  ; ori r11, r11, 0x5F49  -> r11 = 0x42C25F49
    insrdi r3, r11, 32, 0                     -> r3  = 0x42C25F49_85B5C4F4
    bl   Attrib::FindCollectionWithDefault

insrdi with width 32 at position 0 sets the HIGH doubleword, so the callee is
handed the whole 64-bit class key in r3 -- exactly as FindCollection is. Deciding
it discards half needs the callee's own asm, which is not exported (0x82808400
has no per-function JSON), so the claim had nothing behind it.

Corrected to `(u64 luClassKey)` end to end -- declaration, the caller (now passing
the assembled 0x42C25F4985B5C4F4 rather than a truncated int constant), and the
link stub's own signature.

Why it matters even though the body is still a stub: the header records that the
identical truncation on FindCollection was "wrong in a way that silently
mis-resolved EVERY lookup", because the dropped half was the part that
distinguished collections. When G5's body is recovered it will now receive the
key the console passes instead of half of it -- and a wrong-key resolve fails
silently, which is the worst way for this to fail.

Boot verified: zero asserts, green to the junkyard.

+++ src/GameSource/World/WorldLinkStubs.cpp
@@ -2677,7 +2677,7 @@ struct BrnPhysics::Deformation::WheelPhysicalStates & BrnPhysics::Deformation::W
 // r4/r5 untouched. Real in attribinstance.cpp now; the no-arg spelling is retired.)
 
 // LINK STUB (attrib mount closure): gap G5 sibling (Gen:: ChangeWithDefault edit path).
-Attrib::Collection * Attrib::FindCollectionWithDefault(int)
+Attrib::Collection * Attrib::FindCollectionWithDefault(u64)
 {
     CGS_ASSERT(false, "Attrib::FindCollectionWithDefault: attrib gap G5 -- reconstruct");
     return 0;
